+++ bali/application.py
@@ -5,7 +5,7 @@ import uvicorn
 from fastapi import FastAPI
 from starlette.middleware.cors import CORSMiddleware
 
-from .utils import singleton
+from ._utils import singleton
 
 
 @singleton
