+++ requirements.txt
@@ -1,3 +1,4 @@
+dateparser==1.0.0
 fastapi[all]==0.61.1
 grpcio==1.32.0
 grpcio-tools==1.32.0
